Removed .class files, changed paths to relative
.gitignore has *.class in and the .class files already in the repo should
now be removed. The .project file is staying because I'm happy to keep with
Eclipse with this, wherever I'm working on it.

Also, the files that had absolute paths (that don't work on this machine, of
course, given I ported this from another machine) are now relative paths
that work with the way Eclipse calculates them.